move everything into repository root folder
