--- TelegramBot/.idea/.name
@@ -1,1 +1,0 @@
-main.py
