+++ weather.py
@@ -1,0 +1,2 @@
+def get_weather_string():
+    return 'weather will be here'
